+++ api/src/middleware/authenticateToken.ts
@@ -21,12 +21,9 @@ const authenticateToken: RequestHandler = (req, res, next) => {
   return next();
 };
 
+// Authenticates tokens and extracts tokens' payloads
 const authenticateTokens: RequestHandler = async (req, res, next) => {
-  // Get authorization header value
   console.log("authenticateToken running...");
-  // console.log(req.headers);
-  // console.log(req.cookies);
-  // console.log("req.user:", req.user);
 
   const failedResponse = {
     status: "fail",
@@ -41,13 +38,31 @@ const authenticateTokens: RequestHandler = async (req, res, next) => {
     return;
   }
 
-  // Verify access and refresh tokens
-  const { payload, expired } = await verifyJWT(accessToken!);
+  // Verify access token
+  const { payload, expired } = await verifyJWT(accessToken);
+
+  if (payload) {
+    req.user = payload.user;
+    return next();
+  }
+
+  // If the access token is invalid
+  // Verify refresh token
   const { payload: refresh } = await (expired && refreshToken
     ? verifyJWT(refreshToken)
     : Promise.resolve({ payload: null }));
 
+  if (!refresh) {
+    res.status(403).json(failedResponse);
+    return;
+  }
+
+  const { user } = refresh;
+  req.user = user;
+  req.accessToken = null;
+
   return next();
 };
 
-export default authenticateToken;
+// export default authenticateToken;
+export default authenticateTokens;