+++ api/src/routes/auth.route.ts
@@ -4,6 +4,8 @@ import authenticateToken from "../middleware/authenticateToken";
 import { initSchema } from "../validators/query.validator";
 import { validateQuery } from "../validators/validators";
 
+import authenticateTokens from "../middleware/authenticateToken";
+
 const authRoutes = () => {
   const { authorize, refreshAccessToken, login, logout, signup } =
     authController;
@@ -16,7 +18,7 @@ const authRoutes = () => {
   // curl -w "\n" -X POST http://localhost:3001/auth
   authRouter.post("/login", login);
   authRouter.post("/logout", logout);
-  authRouter.post("/refresh", refreshAccessToken);
+  authRouter.post("/refresh", authenticateTokens, refreshAccessToken);
   authRouter.post("/signup", signup);
 
   // curl -w "\n" -X POST http://localhost:3001/pizza/:id